# openapi.yaml
openapi: 3.0.1
info:
  title: Memory AI Mini API
  description: API do zapisu i odczytu danych oraz przesyłania plików do Google Drive.
  version: '1.0'
servers:
  - url: https://twoja-domena.onrender.com
paths:
  /memory/{topic}:
    get:
      summary: Pobierz pamięć dla tematu
      parameters:
        - name: topic
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Zawartość pamięci
    post:
      summary: Zapisz nową notatkę
      parameters:
        - name: topic
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
      responses:
        '200':
          description: Notatka zapisana
  /memory/sync-gdrive:
    post:
      summary: Synchronizuj lokalne pliki do Google Drive
      responses:
        '200':
          description: Lista przesłanych plików
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  uploaded:
                    type: array
                    items:
                      type: string
  /memory/list-gdrive:
    get:
      summary: Wyświetl pliki na Google Drive
      responses:
        '200':
          description: Lista plików
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  files:
                    type: array
                    items:
                      type: object
  /memory/download-gdrive/{file}:
    get:
      summary: Pobierz plik tekstowy z Google Drive
      parameters:
        - name: file
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Treść pliku
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Plik nie znaleziony